Update templatesObject.xml
For most Frames added the following code to inherits="BackdropTemplate"  I am still getting a DropDown error but if I add it to that frame it causes quite a few things to break so I am leaving it off for the moment.  This fixes the settings and edit for lunarsphere although the backgrounds for the frames are transparent.  Haven't figured out how to change their color.  Transparent working panes are better than no panes at all.